requirements.txt
xhtml2pdf
pymongo
pymodm
django-bootstrap-form
djangoajax==2.3.6
